Fix: dynamicAnimatorCompletion block not called
I got some weird issues with my app because `viewDidAppear` was not called sometimes.
Then I saw that the `dynamicAnimatorCompletion` block was not called from:
`- (void)setPaneViewController:(UIViewController *)paneViewController animated:(BOOL)animated completion:(void (^)(void))completion` at line 419.
There is a `[self setPaneState:MSDynamicsDrawerPaneStateClosed animated:animated allowUserInterruption:YES completion:^{` at line 449 that ultimately calls `didMoveToParentViewController`, a really important method for children viewControllers to have.
It turns out that when the central pane appears and you move quickly a `collectionView` or move the central pane, the completionBlock is not called causing a lot of troubles with `navigationController`. I fixed that by calling the block anyway when gesture state is not `UIGestureRecognizerStatePossible`.
I doubted about moving the run of the completionBlock above so I just move it do a method.

Anyway, this fixes a lot of issues I had and I hope it can help others too :) Love this drawer! :+1: